% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_level_visualiser.R
\name{add_levels}
\alias{add_levels}
\title{Add Multiple \code{geom_node_point} Layers for Different Levels of 'Omics Data}
\usage{
add_levels(levels, size = 6, alpha = 1, stroke = 1, size_increment = 3, ...)
}
\arguments{
\item{levels}{A character vector specifying the names of vertex attributes (e.g., 'omics levels') to visualize.}

\item{size}{Initial point size for the topmost level. Default is 6.}

\item{alpha}{Alpha transparency of the points. Default is 1 (fully opaque).}

\item{stroke}{Stroke width of the points. Default is 1.}

\item{size_increment}{Amount to decrease the point size for each successive level. Default is 3.}

\item{...}{Additional aesthetic arguments passed to \code{geom_node_point()}.}
}
\value{
A list of \code{geom_node_point} layers that can be added to a \code{ggraph} plot.
}
\description{
This function creates a list of \code{geom_node_point} layers (for use with \code{ggraph}) to visualize different 'omics-style levels as overlapping node aesthetics.
Each level is rendered with a customizable point size, stroke width, and alpha transparency. Point size decreases for each subsequent level to show multi-layered detail.
}
\examples{
# Assuming tube_graph is an igraph object with vertex attributes:
# EntryTapCount and ExitTapCount
ggraph(tube_graph, layout = layout_df) +
  geom_edge_link(aes(colour = I(color)), show.legend = FALSE) +
  add_levels(levels = c("EntryTapCount", "ExitTapCount"), size = 5, size_increment = 2, alpha = 0.8) +
  theme_graph()

}
