% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_level_visualiser.R
\name{create_combined_legend}
\alias{create_combined_legend}
\title{Create a Combined Legend Grid}
\usage{
create_combined_legend(L_df, levels, colours, nrow = 1, ncol = 2)
}
\arguments{
\item{L_df}{A data frame containing the data to be plotted. It should include columns for each 'omics level specified in the \code{levels} argument.}

\item{levels}{A character vector of column names of levels in \code{L_df} for which legends should be created. Each level should have a corresponding gradient scale.}

\item{colours}{A vector of colours to be used for the gradient scale. This should include at least two colours for a gradient.}

\item{nrow}{Number of rows in the combined legend grid. Default is 1.}

\item{ncol}{Number of columns in the combined legend grid. Default is 2.}
}
\value{
A \code{ggplot} object containing the combined grid of legends.
}
\description{
This function generates a combined grid of legends for multiple omics levels of a given dataset.
Each legend displays a gradient scale corresponding to a specific level.
The resulting legends are combined into a grid layout.
}
\examples{
# Example data frame
L_df <- data.frame(
  mRNA_Factor1 = runif(100, 0, 1),
  protein_Factor1 = runif(100, 0, 1)
)

# Create a combined legend grid
create_combined_legend(
  L_df = L_df,
  levels = c("mRNA_Factor1", "protein_Factor1"),
  colours = c("#000000", "#FFFFFF", "#BA0000"),
  nrow = 1,
  ncol = 2
)

}
