% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_level_visualiser.R
\name{create_custom_legend}
\alias{create_custom_legend}
\title{Create a Custom Legend with Geom Points for Different 'omics Levels}
\usage{
create_custom_legend(
  df,
  levels,
  size = 6,
  size_increment = 3,
  alpha = 1,
  stroke = 1,
  title = "Custom Legend",
  ...
)
}
\arguments{
\item{df}{Data frame containing the data for the legend.}

\item{levels}{A character vector containing the names of factor levels.}

\item{size}{Initial size of the points for the first level. Default is 6.}

\item{size_increment}{Amount to decrease the size for each subsequent level. Default is 3.}

\item{alpha}{Alpha transparency for the points. Default is 1 (opaque).}

\item{stroke}{Stroke width for the points. Default is 1.}

\item{title}{Title for the legend plot. Default is "Custom Legend".}

\item{...}{Additional arguments to be passed to \code{geom_point}.}
}
\value{
A \code{ggplot} object representing the custom legend.
}
\description{
This function creates a custom legend as a single node graph with different
'omics levels represented by varying point sizes. A title and a square border
are added around the node.
}
\examples{
# Create a custom legend with different levels
custom_legend <- create_custom_legend(df = my_data, levels = c("level1", "level2"))

}
