% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualisation.R
\name{londonUnderground_plot}
\alias{londonUnderground_plot}
\title{Plot a Biological Network Using a TFL-style Layout}
\usage{
londonUnderground_plot(
  graph,
  path_layout,
  nudge_label_y = 0.5,
  size = 4,
  stroke = 2
)
}
\arguments{
\item{graph}{An igraph object representing the biological network.}

\item{path_layout}{A layout matrix or data frame.}

\item{nudge_label_y}{Vertical adjustment for node labels.}

\item{size}{Font size for node labels.}

\item{stroke}{Node stroke width.}

\item{color}{A character string or named vector for edge colors.}
}
\value{
A ggplot2 object of the plotted network.
}
\description{
Plot a Biological Network Using a TFL-style Layout
}
