% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualisation.R
\name{theme_tfl}
\alias{theme_tfl}
\title{TFL-style ggplot theme and geoms for biological network plots}
\usage{
theme_tfl(
  nudge_label_y = 0.5,
  node.size = 8,
  stroke = 2,
  text.size = 4,
  show.label = TRUE
)
}
\arguments{
\item{nudge_label_y}{Vertical adjustment for node labels.}

\item{node.size}{Size of the node points.}

\item{stroke}{Stroke width of the nodes.}

\item{text.size}{Font size for node labels.}

\item{show.label}{Logical; whether to display node labels.}

\item{color}{Optional: A character string or named vector for edge colors.}
}
\value{
A list of ggplot2 layers styled in a London Underground aesthetic.
}
\description{
TFL-style ggplot theme and geoms for biological network plots
}
