% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_level_visualiser.R
\name{add_levels}
\alias{add_levels}
\title{Add Multiple \code{geom_point} Layers for Different Levels of 'omics data}
\usage{
add_levels(df, x, y, size = 6, alpha = 1, stroke = 1, size_increment = 3, ...)
}
\arguments{
\item{df}{Data frame containing the 'omics data for the plot.}

\item{x}{Name of the column in \code{df} to be used for the x-axis.}

\item{y}{Name of the column in \code{df} to be used for the y-axis.}

\item{size}{Initial size of the points for the first level. Default is 6.}

\item{alpha}{Alpha transparency for the points. Default is 1 (opaque).}

\item{stroke}{Stroke width for the points. Default is 1.}

\item{size_increment}{Amount to decrease the size for each subsequent level. Default is 3.}

\item{...}{Additional arguments to be passed to \code{geom_point}.}

\item{levels}{A character vector containing the names of 'omics levels.}
}
\value{
A list of \code{geom_point} layers.
}
\description{
This function creates a list of \code{geom_point} layers for different 'omics levels
with customizable sizes, alpha transparency, and stroke width. The size of the
points decreases incrementally for each subsequent level.
}
\examples{
# Create a list of layers with different levels
layers <- add_levels(df = my_data, x = "x", y = "y", levels = c("level1", "level2"))
}
