% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualisation.R
\name{create_custom_layout}
\alias{create_custom_layout}
\title{Create a Custom Layout for Nodes Based on Their Type}
\usage{
create_custom_layout(graph, type_labels, y_position)
}
\arguments{
\item{graph}{An igraph object.}

\item{type_labels}{A vector of node types to include in the layout.}

\item{y_position}{Numeric value specifying the vertical position for the nodes of this type.}
}
\value{
A data frame with the x and y coordinates of the nodes.
}
\description{
Create a Custom Layout for Nodes Based on Their Type
}
