% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualisation.R
\name{londonUnderground_plot}
\alias{londonUnderground_plot}
\title{Plot a Biological Network Using a Custom Layout}
\usage{
londonUnderground_plot(graph, path_layout, color)
}
\arguments{
\item{graph}{An igraph object representing the biological network.}

\item{path_layout}{A matrix or data frame representing the layout of the nodes.}

\item{color}{A character string specifying the color for edges.}
}
\value{
A ggplot2 object of the plotted network.
}
\description{
Plot a Biological Network Using a Custom Layout
}
